fix: split archive deploy manifest config
