+++ src/core/extension/manifest.rs
@@ -14,13 +14,13 @@ pub use super::manifest_action_config::{
 };
 pub use super::manifest_config::{
     AutofixVerifyConfig, BenchConfig, BuildConfig, CliAutoFlag, CliAutoFlagCondition, CliConfig,
-    CliHelpConfig, DatabaseCliConfig, DatabaseConfig, DeployArchiveInstallPolicy, DeployOverride,
-    DeployRequiredHeader, DeployVerification, DepsConfig, DiscoveryConfig, EnvProviderConfig,
-    FileContainsCondition, LintChangedFileRoute, LintConfig, RemotePathInferenceRule,
-    RemotePathRootRule, RequirementsConfig, SinceTagConfig, TestChangedFileExclusiveEnv,
-    TestChangedFileRouting, TestChangedFileRoutingStrategy, TestConfig, TraceConfig,
-    VersionPatternConfig,
+    CliHelpConfig, DatabaseCliConfig, DatabaseConfig, DeployOverride, DeployVerification,
+    DepsConfig, DiscoveryConfig, EnvProviderConfig, FileContainsCondition, LintChangedFileRoute,
+    LintConfig, RemotePathInferenceRule, RemotePathRootRule, RequirementsConfig, SinceTagConfig,
+    TestChangedFileExclusiveEnv, TestChangedFileRouting, TestChangedFileRoutingStrategy,
+    TestConfig, TraceConfig, VersionPatternConfig,
 };
+pub use super::manifest_deploy_config::{DeployArchiveInstallPolicy, DeployRequiredHeader};
 pub use super::manifest_sidecar::{StructuredSidecarContract, StructuredSidecarDeclaration};
 
 /// Type of action that can be executed by a extension.
